+++ src/main/java/com/twilio/rest/api/v2010/account/usage/record/AllTime.java
@@ -163,6 +163,7 @@ public class AllTime extends Resource {
         PHONENUMBERS_TOLLFREE("phonenumbers-tollfree"),
         PREMIUMSUPPORT("premiumsupport"),
         PROXY("proxy"),
+        PROXY_ACTIVE_SESSIONS("proxy-active-sessions"),
         PV("pv"),
         PV_COMPOSITION_MEDIA_DOWNLOADED("pv-composition-media-downloaded"),
         PV_COMPOSITION_MEDIA_ENCRYPTED("pv-composition-media-encrypted"),