+++ src/main/java/com/twilio/rest/preview/understand/assistant/intent/IntentActionsUpdater.java
@@ -5,7 +5,7 @@
  *       /       /
  */
 
-package com.twilio.rest.preview.understand.assistant;
+package com.twilio.rest.preview.understand.assistant.intent;
 
 import com.twilio.base.Updater;
 import com.twilio.converter.Converter;
@@ -25,50 +25,21 @@ import java.util.Map;
  * change. Use them with caution. If you currently do not have developer preview
  * access, please contact help@twilio.com.
  */
-public class IntentUpdater extends Updater<Intent> {
+public class IntentActionsUpdater extends Updater<IntentActions> {
     private final String pathAssistantSid;
-    private final String pathSid;
-    private String friendlyName;
-    private String uniqueName;
+    private final String pathIntentSid;
     private Map<String, Object> actions;
 
     /**
-     * Construct a new IntentUpdater.
+     * Construct a new IntentActionsUpdater.
      * 
      * @param pathAssistantSid The assistant_sid
-     * @param pathSid The sid
+     * @param pathIntentSid The intent_sid
      */
-    public IntentUpdater(final String pathAssistantSid, 
-                         final String pathSid) {
+    public IntentActionsUpdater(final String pathAssistantSid, 
+                                final String pathIntentSid) {
         this.pathAssistantSid = pathAssistantSid;
-        this.pathSid = pathSid;
-    }
-
-    /**
-     * A user-provided string that identifies this resource. It is non-unique and
-     * can up to 255 characters long..
-     * 
-     * @param friendlyName A user-provided string that identifies this resource. It
-     *                     is non-unique and can up to 255 characters long.
-     * @return this
-     */
-    public IntentUpdater setFriendlyName(final String friendlyName) {
-        this.friendlyName = friendlyName;
-        return this;
-    }
-
-    /**
-     * A user-provided string that uniquely identifies this resource as an
-     * alternative to the sid. Unique up to 64 characters long..
-     * 
-     * @param uniqueName A user-provided string that uniquely identifies this
-     *                   resource as an alternative to the sid. Unique up to 64
-     *                   characters long.
-     * @return this
-     */
-    public IntentUpdater setUniqueName(final String uniqueName) {
-        this.uniqueName = uniqueName;
-        return this;
+        this.pathIntentSid = pathIntentSid;
     }
 
     /**
@@ -77,7 +48,7 @@ public class IntentUpdater extends Updater<Intent> {
      * @param actions The actions
      * @return this
      */
-    public IntentUpdater setActions(final Map<String, Object> actions) {
+    public IntentActionsUpdater setActions(final Map<String, Object> actions) {
         this.actions = actions;
         return this;
     }
@@ -86,15 +57,15 @@ public class IntentUpdater extends Updater<Intent> {
      * Make the request to the Twilio API to perform the update.
      * 
      * @param client TwilioRestClient with which to make the request
-     * @return Updated Intent
+     * @return Updated IntentActions
      */
     @Override
     @SuppressWarnings("checkstyle:linelength")
-    public Intent update(final TwilioRestClient client) {
+    public IntentActions update(final TwilioRestClient client) {
         Request request = new Request(
             HttpMethod.POST,
             Domains.PREVIEW.toString(),
-            "/understand/Assistants/" + this.pathAssistantSid + "/Intents/" + this.pathSid + "",
+            "/understand/Assistants/" + this.pathAssistantSid + "/Intents/" + this.pathIntentSid + "/Actions",
             client.getRegion()
         );
 
@@ -102,7 +73,7 @@ public class IntentUpdater extends Updater<Intent> {
         Response response = client.request(request);
 
         if (response == null) {
-            throw new ApiConnectionException("Intent update failed: Unable to connect to server");
+            throw new ApiConnectionException("IntentActions update failed: Unable to connect to server");
         } else if (!TwilioRestClient.SUCCESS.apply(response.getStatusCode())) {
             RestException restException = RestException.fromJson(response.getStream(), client.getObjectMapper());
             if (restException == null) {
@@ -118,7 +89,7 @@ public class IntentUpdater extends Updater<Intent> {
             );
         }
 
-        return Intent.fromJson(response.getStream(), client.getObjectMapper());
+        return IntentActions.fromJson(response.getStream(), client.getObjectMapper());
     }
 
     /**
@@ -127,14 +98,6 @@ public class IntentUpdater extends Updater<Intent> {
      * @param request Request to add post params to
      */
     private void addPostParams(final Request request) {
-        if (friendlyName != null) {
-            request.addPostParam("FriendlyName", friendlyName);
-        }
-
-        if (uniqueName != null) {
-            request.addPostParam("UniqueName", uniqueName);
-        }
-
         if (actions != null) {
             request.addPostParam("Actions", Converter.mapToJson(actions));
         }